refactor: `system_profiler` => `system+profiler_memory`
