+++ src-tauri/src/infrastructure/providers/macos/mod.rs
@@ -3,5 +3,5 @@ pub mod gpu_info;
 pub mod io_kit;
 pub mod net_sys;
 pub mod sysctl;
-pub mod system_profiler;
 pub mod system_profiler_displays;
+pub mod system_profiler_memory;
