+++ src-tauri/src/infrastructure/providers/macos/system_profiler_memory.rs
